+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.1.2-beta
 
 * Added `ubuntu-drivers-common` to the default app list to simplify the driver installation process.
+* Removed `apport` and `python3-systemd` to reduce the size of the ISO file.
 
 ## v0.1.1-beta